overwrite and threads

+++ bin/whippet-motif.jl
@@ -90,7 +90,8 @@ function score_motifs( ifile::String,
                        rbps::Vector{RNABindNSeq},
                        genome::Dict{String,ReferenceSequence},
                        rbpspan::Int,
-                       gpsi::Bool=false )
+                       gpsi::Bool=false,
+                       overwrite::Bool=false )
 
    header    = true
    skipfive  = false
@@ -98,6 +99,9 @@ function score_motifs( ifile::String,
 
    used = Dict{String,Bool}()
 
+   if isfile(ofile) && !overwrite
+      return
+   end
    bufout = open(ofile, "w") |> ZlibDeflateOutputStream
 
    for l in eachline( bufferedinput(ifile) )
@@ -245,7 +249,7 @@ function main()
    ofiles = map(x->splitext(basename(x))[1] * ".motif.gz", sfiles)
    fexten = map(x->(split(x, '.')[end] == "gz" ? split(x, '.')[end-1] : split(x, '.')[end]), sfiles)
 
-   for i in 1:length(ifiles)
+   Threads.@threads for i in 1:length(ifiles)
       println(stderr, "Fetching motifs for $(ifiles[i])")
       score_motifs( ifiles[i], ofiles[i], five_mat, three_mat, rbps, genome, rbpspan, fexten[i] == "gpsi" )
    end
